getList() method implemented
Implemented getList() method for the Restaurants class which returns a deep copy of the current Restaurants object in the form of an ArrayList<Restaurants>. The copy contains all the Restaurants in sorted order, with the most recent inspection added to its Inspections object. Regression test for Restaurants class was updated to include testing for the new method. Regression tests for Inspections and Inspection class also updated to new changes made to the classes.