Checking for new updates
Added functionality to check if 20 hours have at least passed since the previous update. Added an alert dialog which allows the user to choose if they would like to update the data. If update is selected by the user, a progress dialog is displayed and the data is being fetched in the background. Once the data has been fetched, the old Restaurants object is discarded and a new one is initialized with the latest date and it displayed in the list view. The progress dialog dismissed once the data is displayed in the list view. RetriveData class was moved to MainActivity to make post and pre execution tasks easier to control.