Initializing with data from Iteration 1
Initialized data set in app with data set provided in iteration 1. If the app is installed for the first time (or uninstalled and reinstalled), the data displayed will be that from iteration 1. If the user does not choose to update the data, then they will have access to only this initial data set. When the app is started back up, user will be offered an update once again.